+++ src/Service/DateTimeHelper.php
@@ -8,6 +8,17 @@ class DateTimeHelper
     ) {
     }
 
+    /**
+     * Returns the first and last date of a given week in a year (ISO 8601).
+     *
+     * @param int $weekNumber The week number.
+     *
+     * @param int|null $year The year. If not provided, the current year will be used.
+     *
+     * @param string $format The date format to be returned. Defaults to 'Y-m-d H:i:s'.
+     *
+     * @return array An array with the first and last date of the week.
+     */
     public function getFirstAndLastDateOfWeek(int $weekNumber, int $year = null, string $format = 'Y-m-d H:i:s'): array
     {
         if (!$year) {
@@ -25,6 +36,17 @@ class DateTimeHelper
         return ['first' => $firstDate, 'last' => $lastDate];
     }
 
+    /**
+     * Returns the first and last date of the specified month and year.
+     *
+     * @param int $monthNumber The month number (1-12).
+     *
+     * @param int|null $year The year. If null, the current year will be used.
+     *
+     * @param string $format The format to use for the returned dates.
+     *
+     * @return array An array containing the first and last date of the specified month and year.
+     */
     public function getFirstAndLastDateOfMonth(int $monthNumber, int $year = null, string $format = 'Y-m-d H:i:s'): array
     {
         if (!$year) {
@@ -43,9 +65,11 @@ class DateTimeHelper
     }
 
     /**
-     * Returns an array of the weeks for the current year (ISO 8601).
+     * Retrieves an array of week numbers for a given year.
      *
-     * @return array
+     * @param int|null $year The year for which to retrieve the week numbers. If null, the current year is used.
+     *
+     * @return array An array of week numbers.
      */
     public function getWeeksOfYear(int $year = null): array
     {
@@ -73,17 +97,28 @@ class DateTimeHelper
         return $weekArray;
     }
 
+    /**
+     * Retrieves an array of months and their corresponding numeric representation of a year.
+     *
+     * @return array An array where the keys are month names and the values are their corresponding numeric representation (1-12).
+     */
     public function getMonthsOfYear(): array
     {
         $months = [];
         for ($i = 1; $i <= 12; ++$i) {
-            $monthName = \DateTime::createFromFormat('!m', (string) $i)->format('F');
-            $months[$monthName] = $i;
+            $months[] = $i;
         }
 
         return $months;
     }
 
+    /**
+     * Retrieves the name of the month for a given month number.
+     *
+     * @param int $monthNumber The month number for which to retrieve the month name.
+     *
+     * @return string The name of the month.
+     */
     public function getMonthName(int $monthNumber): string
     {
         return \DateTime::createFromFormat('!m', (string) $monthNumber)->format('F');