+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [PR-117](https://github.com/itk-dev/economics/pull/117)
   1211: Added hour report
 * [PR-124](https://github.com/itk-dev/economics/pull/124)
-  710: Workload report
+  710: Added workload report
 
 ## [2.3.0] - 2024-06-03